+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -6,7 +6,8 @@ import classNames from 'classnames';
 import { endsWith } from 'lodash';
 import List from '../List';
 import getPluginViewers from '../getPluginViewers';
-import { getTextDirection, kebabToCamelObjectKeys } from './textUtils';
+import { kebabToCamelObjectKeys } from './textUtils';
+import { getTextDirection } from './textDirection';
 import { staticInlineStyleMapper } from '../staticInlineStyleMapper';
 import { combineMappers } from './combineMappers';
 import { getInteractionWrapper, DefaultInteractionWrapper } from './getInteractionWrapper';
